Module async_cell::unsync

Types for single-threaded and no_std use.

Structs

AsyncCell

An async primitive. Similar to a Cell<Option<T>> but awaitable.

Get

Used to clone the value of a cell. Can be constructed directly or with AsyncCell::get.

GetWeak

A similar future to Get, but resolves to None when the cell is dropped. Can be constructed directly or with AsyncCell::get_weak.

GuardedCell

A utility wrapper to set a given AsyncCell<T> when dropped.

Take

Used to remove the value of a cell. Can be constructed directly or with AsyncCell::take.

TakeWeak

A similar future to Take, but resolves to None when the cell is dropped. Can be constructed directly or with AsyncCell::take_weak.